refacto(Repository): separate EntityRepository et PaginatedEntityRepository
