+++ LivlReviewsApi/Data/Category.cs
@@ -1,0 +1,6 @@
+namespace LivlReviewsApi.Data;
+
+public class Category
+{
+    
+}
